# Improve string gather performance