--- a/cpp/benchmarks/string/replace_benchmark.cpp
+++ b/cpp/benchmarks/string/replace_benchmark.cpp
@@ -26,6 +26,8 @@
 
 #include <limits>
 
+#include "string_bench_args.hpp"
+
 class StringReplace : public cudf::benchmark {
 };
 
@@ -69,15 +71,7 @@ static void generate_bench_args(benchmark::internal::Benchmark* b)
   int const min_rowlen = 1 << 5;
   int const max_rowlen = 1 << 13;
   int const len_mult   = 4;
-  for (int row_count = min_rows; row_count <= max_rows; row_count *= row_mult) {
-    for (int rowlen = min_rowlen; rowlen <= max_rowlen; rowlen *= len_mult) {
-      // avoid generating combinations that exceed the cudf column limit
-      size_t total_chars = static_cast<size_t>(row_count) * rowlen;
-      if (total_chars < std::numeric_limits<cudf::size_type>::max()) {
-        b->Args({row_count, rowlen});
-      }
-    }
-  }
+  generate_string_bench_args(b, min_rows, max_rows, row_mult, min_rowlen, max_rowlen, len_mult);
 }
 
 #define STRINGS_BENCHMARK_DEFINE(name)                              \