--- a/cpp/tests/copying/gather_str_tests.cu
+++ b/cpp/tests/copying/gather_str_tests.cu
@@ -47,6 +47,35 @@ TEST_F(GatherTestStr, StringColumn)
   CUDF_TEST_EXPECT_TABLES_EQUAL(expected, got->view());
 }
 
+TEST_F(GatherTestStr, GatherSlicedStringsColumn)
+{
+  cudf::test::strings_column_wrapper strings{{"This", "is", "not", "a", "string", "type"},
+                                             {1, 1, 1, 1, 1, 0}};
+  std::vector<cudf::size_type> slice_indices{0, 2, 2, 3, 3, 6};
+  auto sliced_strings = cudf::slice(strings, slice_indices);
+  {
+    cudf::test::fixed_width_column_wrapper<int16_t> gather_map{{1, 0, 1}};
+    cudf::test::strings_column_wrapper expected_strings{{"is", "This", "is"}, {1, 1, 1}};
+    cudf::table_view expected{{expected_strings}};
+    auto result = cudf::gather(cudf::table_view{{sliced_strings[0]}}, gather_map);
+    CUDF_TEST_EXPECT_TABLES_EQUAL(expected, result->view());
+  }
+  {
+    cudf::test::fixed_width_column_wrapper<int16_t> gather_map{{0, 0, 0}};
+    cudf::test::strings_column_wrapper expected_strings{{"not", "not", "not"}, {1, 1, 1}};
+    cudf::table_view expected{{expected_strings}};
+    auto result = cudf::gather(cudf::table_view{{sliced_strings[1]}}, gather_map);
+    CUDF_TEST_EXPECT_TABLES_EQUAL(expected, result->view());
+  }
+  {
+    cudf::test::fixed_width_column_wrapper<int16_t> gather_map{{2, 1, 0}};
+    cudf::test::strings_column_wrapper expected_strings{{"", "string", "a"}, {0, 1, 1}};
+    cudf::table_view expected{{expected_strings}};
+    auto result = cudf::gather(cudf::table_view{{sliced_strings[2]}}, gather_map);
+    CUDF_TEST_EXPECT_TABLES_EQUAL(expected, result->view());
+  }
+}
+
 TEST_F(GatherTestStr, Gather)
 {
   std::vector<const char*> h_strings{"eee", "bb", "", "aa", "bbb", "ééé"};
@@ -98,7 +127,7 @@ TEST_F(GatherTestStr, GatherDontCheckOutOfBounds)
   CUDF_TEST_EXPECT_COLUMNS_EQUAL(results->view().column(0), expected);
 }
 
-TEST_F(GatherTestStr, GatherZeroSizeStringsColumn)
+TEST_F(GatherTestStr, GatherEmptyMapStringsColumn)
 {
   cudf::column_view zero_size_strings_column(
     cudf::data_type{cudf::type_id::STRING}, 0, nullptr, nullptr, 0);
@@ -108,4 +137,19 @@ TEST_F(GatherTestStr, GatherZeroSizeStringsColumn)
                                       gather_map.end(),
                                       cudf::out_of_bounds_policy::NULLIFY);
   cudf::test::expect_strings_empty(results->get_column(0).view());
+}
+
+TEST_F(GatherTestStr, GatherZeroSizeStringsColumn)
+{
+  cudf::column_view zero_size_strings_column(
+    cudf::data_type{cudf::type_id::STRING}, 0, nullptr, nullptr, 0);
+  std::vector<int32_t> h_map{0};
+  cudf::test::fixed_width_column_wrapper<int32_t> gather_map(h_map.begin(), h_map.end());
+  cudf::column_view gather_view = gather_map;
+  cudf::test::strings_column_wrapper expected{std::pair<std::string, bool>{"", false}};
+  auto results = cudf::detail::gather(cudf::table_view({zero_size_strings_column}),
+                                      gather_view.begin<int32_t>(),
+                                      gather_view.end<int32_t>(),
+                                      cudf::out_of_bounds_policy::NULLIFY);
+  CUDF_TEST_EXPECT_COLUMNS_EQUAL(expected, results->get_column(0).view());
 }