This improves string gather performance by using an algorithm with character-level parallelism.  A string gather benchmark with varying row counts and row lengths has also been added.