@jlowe 2021-02-24
Benchmark numbers.
Before:
```
StringCopy/gather/4096/32/manual_time           0.108 ms        0.128 ms         5990 bytes_per_second=639.892M/s
StringCopy/gather/4096/128/manual_time          0.155 ms        0.175 ms         4285 bytes_per_second=1.70927G/s
StringCopy/gather/4096/512/manual_time          0.378 ms        0.396 ms         1842 bytes_per_second=2.83254G/s
StringCopy/gather/4096/2048/manual_time          1.22 ms         1.24 ms          568 bytes_per_second=3.49513G/s
StringCopy/gather/4096/8192/manual_time          4.59 ms         4.61 ms          153 bytes_per_second=3.69724G/s
StringCopy/gather/32768/32/manual_time          0.121 ms        0.140 ms         5509 bytes_per_second=4.46532G/s
StringCopy/gather/32768/128/manual_time         0.203 ms        0.221 ms         3371 bytes_per_second=10.503G/s
StringCopy/gather/32768/512/manual_time         0.574 ms        0.597 ms         1162 bytes_per_second=14.8195G/s
StringCopy/gather/32768/2048/manual_time         1.97 ms         1.99 ms          355 bytes_per_second=17.3396G/s
StringCopy/gather/32768/8192/manual_time         7.66 ms         7.68 ms           91 bytes_per_second=17.7785G/s
StringCopy/gather/262144/32/manual_time         0.255 ms        0.273 ms         2727 bytes_per_second=16.9522G/s
StringCopy/gather/262144/128/manual_time         4.15 ms         4.17 ms          169 bytes_per_second=4.11681G/s
StringCopy/gather/262144/512/manual_time         36.9 ms         36.9 ms           19 bytes_per_second=1.84783G/s
StringCopy/gather/262144/2048/manual_time         188 ms          188 ms            4 bytes_per_second=1.44939G/s
StringCopy/gather/2097152/32/manual_time         7.21 ms         7.23 ms           97 bytes_per_second=4.78411G/s
StringCopy/gather/2097152/128/manual_time        43.9 ms         44.0 ms           16 bytes_per_second=3.10625G/s
StringCopy/gather/2097152/512/manual_time         309 ms          309 ms            2 bytes_per_second=1.76693G/s
StringCopy/gather/16777216/32/manual_time        76.2 ms         76.2 ms            9 bytes_per_second=3.62227G/s
```
After:
```
-----------------------------------------------------------------------------------------------------
Benchmark                                           Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------------
StringCopy/gather/4096/32/manual_time           0.096 ms        0.116 ms         6959 bytes_per_second=723.264M/s
StringCopy/gather/4096/128/manual_time          0.101 ms        0.120 ms         6198 bytes_per_second=2.62843G/s
StringCopy/gather/4096/512/manual_time          0.129 ms        0.148 ms         4996 bytes_per_second=8.27184G/s
StringCopy/gather/4096/2048/manual_time         0.234 ms        0.252 ms         2927 bytes_per_second=18.2411G/s
StringCopy/gather/4096/8192/manual_time         0.651 ms        0.671 ms         1040 bytes_per_second=26.0267G/s
StringCopy/gather/32768/32/manual_time          0.121 ms        0.140 ms         5523 bytes_per_second=4.44146G/s
StringCopy/gather/32768/128/manual_time         0.182 ms        0.200 ms         3737 bytes_per_second=11.7394G/s
StringCopy/gather/32768/512/manual_time         0.423 ms        0.442 ms         1650 bytes_per_second=20.1458G/s
StringCopy/gather/32768/2048/manual_time         1.40 ms         1.42 ms          496 bytes_per_second=24.3906G/s
StringCopy/gather/32768/8192/manual_time         5.28 ms         5.30 ms          131 bytes_per_second=25.7635G/s
StringCopy/gather/262144/32/manual_time         0.309 ms        0.328 ms         2245 bytes_per_second=13.9549G/s
StringCopy/gather/262144/128/manual_time        0.860 ms        0.879 ms          805 bytes_per_second=19.8374G/s
StringCopy/gather/262144/512/manual_time         3.07 ms         3.09 ms          228 bytes_per_second=22.194G/s
StringCopy/gather/262144/2048/manual_time        11.9 ms         12.0 ms           59 bytes_per_second=22.8614G/s
StringCopy/gather/2097152/32/manual_time         2.04 ms         2.05 ms          344 bytes_per_second=16.9527G/s
StringCopy/gather/2097152/128/manual_time        6.97 ms         6.99 ms          100 bytes_per_second=19.5739G/s
StringCopy/gather/2097152/512/manual_time        26.8 ms         26.8 ms           26 bytes_per_second=20.3467G/s
StringCopy/gather/16777216/32/manual_time        18.9 ms         18.9 ms           37 bytes_per_second=14.6127G/s
```