@kkraus14 2021-02-24
> Very nice improvement. There's room for even further improvement by doing larger than 1B read/writes when possible as well, but that would be quite a bit more complex and I'm guessing the current speedup is satisfactory.

I think this is something that @gaohao95 is planning on looking into in the near future.