@ttnghia 2021-02-26
Great work. But since your work in mostly related to `include/cudf/strings/*.*` and `src/strings/*.*`,  should we change `cpp/tests/copying/gather_str_tests.cu` into `cpp/tests/strings/gather_str_tests.cu`?