@jlowe 2021-02-26
> should we change cpp/tests/copying/gather_str_tests.cu into cpp/tests/strings/gather_str_tests.cu?

This test is organized by how the public API is organized.  `cudf::gather` is in `copying.hpp` so IMO it makes sense that the tests for it are under `copying`.  If we want to shuffle test files around, I think that's better left to a PR dedicated to that.